feat: 관리자계정수정

+++ src/main/java/com/nhnacademy/inkbridge/backend/storage/admin/AdminAccountJpaRepository.java
@@ -2,5 +2,5 @@ package com.nhnacademy.inkbridge.backend.storage.admin;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AdminAccountJpaRepository extends JpaRepository<AdminEntity, Long> {
+public interface AdminAccountJpaRepository extends JpaRepository<AdminEntity, Integer> {
 }
